- mudanca em telemetry_module.h
- variaveis globais viraram declaracoes com extern
-  funcoes passaram a ser apenas prototipos
- criei telemetry_module.cpp com:
- definicoes unicas de packet_count e previous_millis
- implementacao de getdatastring, printBoth e logData
